Decompose Dialog: extract diff editor
